大模型微调肺炎检测/requirements.txt
torch>=2.1.0
transformers>=4.45.0
peft>=0.12.0
trl>=0.12.0
accelerate>=0.34.0
qwen-vl-utils>=0.0.8
pillow>=10.0.0
matplotlib>=3.7.0
numpy>=1.24.0
